requirements.txt
absl-py==2.1.0
accelerate
tracr @ git+https://github.com/deepmind/tracr
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
apache-libcloud==3.8.0
appdirs==1.4.4
argcomplete==3.2.3
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
boto3==1.34.74
botocore==1.34.74
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
chex==0.1.86
circuitbreaker==1.4.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.25.0
coloredlogs==15.0.1
colorspacious==1.1.2
composer
contourpy==1.2.0
coolname==2.2.0
cramjam==2.8.3
cryptography==42.0.5
cycler==0.12.1
datasets
decorator==5.1.1
deepspeed==0.14.0
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
dm-haiku==0.0.12
docker==7.0.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.5.0
entrypoints==0.4
etils==1.7.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
fire==0.6.0
Flask==3.0.2
flatbuffers==24.3.25
flax==0.8.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.4.0
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jmp==0.0.4
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lightning-utilities==0.11.2
lit==15.0.7
llm-foundry
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mlflow
mosaicml
mosaicml-cli
mosaicml-streaming
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oci==2.125.0
omegaconf==2.3.0
onnx
onnxruntime
openai==1.21.1
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.7
packaging==22.0
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pathtools
pexpect==4.9.0
pickleshare==0.7.5
pillow
pipreqs==0.5.0
platformdirs==4.2.2
portalocker==2.8.2
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pynvml==11.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-snappy==0.7.1
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
querystring-parser==1.2.4
questionary==2.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.28.1
responses==0.18.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safetensors
scikit-learn
scipy
seaborn==0.13.2
sentencepiece==0.1.97
sentry-sdk==1.44.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
slack_sdk==3.27.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorstore==0.1.56
termcolor==2.4.0
threadpoolctl==3.4.0
tinycss2==1.3.0
tokenizers==0.15.2
toolz==0.12.1
torch
torch-optimizer==0.3.0
torchaudio
torchmetrics
torchvision
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers
triton==2.1.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
validators==0.24.0
wandb==0.15.12
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1
zstd==1.5.5.1
